move cugraph tests to  tests/cugraph/test_basics.py
